requirements.txt
PyAutoGUI==0.9.54
Pillow==11.2.1
opencv-python==4.11.0.86